fix most tests
