+++ tests/api/test_api_site.py
@@ -1,4 +1,5 @@
 from plone import api
+from plone.distribution.api import distribution as dist_api
 from plone.distribution.api import site as site_api
 from Products.CMFPlone.Portal import PloneSite
 
@@ -53,3 +54,14 @@ class TestApiSite:
         site = sites[-1]
         assert isinstance(site, PloneSite)
         assert site.title == new_site.title
+
+    def test_get_creation_report_new_site(self, site):
+        from datetime import datetime
+        from plone.distribution.core import SiteCreationReport
+
+        report = dist_api.get_creation_report(site)
+        assert isinstance(report, SiteCreationReport)
+        assert report.name == "default"
+        assert isinstance(report.date, datetime)
+        assert isinstance(report.answers, dict)
+        assert report.answers["title"] == site.title
